sync code

+++ src/common/backend/utils/fmgr/fmgr.cpp
@@ -2515,7 +2515,7 @@ struct varlena* pg_detoast_datum_copy(struct varlena* datum)
     }
 }
 
-struct varlena* pg_detoast_datum_slice(struct varlena* datum, int32 first, int32 count)
+struct varlena* pg_detoast_datum_slice(struct varlena* datum, int64 first, int32 count)
 {
     /* Only get the specified portion from the toast rel */
     return heap_tuple_untoast_attr_slice(datum, first, count);
